First crack at eXO CatalogFrontend implementation
CatalogFrontend instantiates a PastryNode and a Jetty server object.
At this stage, we have handlers for GetUserProfile and SetUserProfile.
The former acts on the local node only and thus returns immediately,
while the latter makes use of FreePastry continuations and thus needs to
use the intermediate request/response queue. All I/O with the web
application is in JSON.

Also adding Apache Wicket bindings for now.