More revamping of frontend handlers
Moved some more stuff into the AbstractHandler base class.
Cleaned-up the interface of sub-handlers with the base class; no child
has now direct access to the shared queue, or cares about the result
format. Sub-handlers can now make use of the sendStatus(...) and
queueStatus(...) methods for this cause. Further introduced a
RequestStatus enumeration that can be used along with these methods, to
make this new interface a bit stricter.