Removing htmlparsing package
The idea of parsing HTML and text documents to compute term frequencies
and utterly index them on the network is interesting. However, it may
result in massively excessive indexing, as each document may thus
contain a large number of terms (and I didn't like that htmlparsing
package anyway). Removing this for now until we have a more
IR-acceptable approach of dealing with stopwords and the like.