DHTService is no more!
DHTService was a mere copy of PastImpl, plus a couple home-brewed
lookup() variants. Keeping it around made no sense, especially as we
don't have the time to backport changesets and bugfixes from the FP repo
to our copy. This diff moves the extra functions into CatalogService and
makes it a direct descendant of PastImpl.